Add base loaded section for detail
