+++ feature/home/detail/src/main/kotlin/jp/shirataki707/yamato/feature/home/detail/ui/section/DetailLoadedSection.kt
@@ -1,2 +1,12 @@
 package jp.shirataki707.yamato.feature.home.detail.ui.section
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+internal fun DetailLoadedSection(
+    sectionState: DetailLoadedSectionState,
+    modifier: Modifier = Modifier,
+) {
+    // TODO: Implement the loaded section
+}
