+++ feature/home/main/src/main/kotlin/jp/shirataki707/yamato/feature/home/main/ui/section/HomeLoadedSectionState.kt
@@ -7,6 +7,7 @@ import android.net.Uri
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalContext
+import jp.shirataki707.yamato.core.common.utils.openYoutubeVideo
 import jp.shirataki707.yamato.core.model.data.DetailPageConfig
 import jp.shirataki707.yamato.core.model.data.VideoResources
 
@@ -39,19 +40,5 @@ internal fun rememberHomeLoadedSectionState(
             onVideoClick = onVideoClick,
             onMoreButtonClick = onMoreButtonClick,
         )
-    }
-}
-
-private fun openYoutubeVideo(context: Context, videoId: String) {
-    val appUri = Uri.parse("vnd.youtube:$videoId")
-    val webUri = Uri.parse("https://www.youtube.com/watch?v=$videoId")
-
-    val appIntent = Intent(Intent.ACTION_VIEW, appUri)
-    val webIntent = Intent(Intent.ACTION_VIEW, webUri)
-
-    try {
-        context.startActivity(appIntent)
-    } catch (e: ActivityNotFoundException) {
-        context.startActivity(webIntent)
     }
 }
